Conflicts:
	ProjectSDL2.1/ProjectSDL2.1.vcxproj
	ProjectSDL2.1/ProjectSDL2.1.vcxproj.filters
	ProjectSDL2.1/Scene.cpp

+++ ProjectSDL2.1/GLCollisionHandler.cpp
@@ -12,7 +12,7 @@ void GLCollisionHandler::CheckCollisions(float deltaTime)
 	glm::vec3 distance;
 	float distSqrd;
 	AABB wall(glm::vec3(0), glm::vec3(125, 48, 86));
-	for (int i = 0; i < players.size(); i++)
+	for (size_t i = 0; i < players.size(); i++)
 	{
 		distance = players.at(i)->GetTransform().GetPos() - players.at(1 - i)->GetTransform().GetPos();
 		distSqrd = glm::dot(distance,distance);
@@ -73,11 +73,11 @@ void GLCollisionHandler::CheckCollisions(float deltaTime)
 			players.at(i)->getVelocity() -= normal * glm::dot(players.at(i)->GetVelocity(), normal);
 		}
 
-		for (int j = 0; j < players.at(i)->GetProjectiles().size(); j++)
+		for (size_t j = 0; j < players.at(i)->GetProjectiles().size(); j++)
 		{
 			if (players.at(i)->GetProjectiles().at(j)->GetBoundingBox().containsAABB(players.at(1 - i)->GetBoundingBox()))
 			{
-				players.at(1 - i)->HandleCollision(GLPlayer::EATING, deltaTime, glm::vec3(-0.10));
+				players.at(1 - i)->HandleCollision(GLPlayer::EATING, deltaTime, glm::vec3(-0.10f));
 				players.at(1 - i)->HandleCollision(GLPlayer::HIT, deltaTime,players.at(i)->GetProjectiles().at(j)->GetForward() * 30.0f);
 				players.at(i)->GetProjectiles().at(j)->Inactivate();
 			}
@@ -132,7 +132,7 @@ void GLCollisionHandler::AddPlayer(GLPlayer * player)
 
 void GLCollisionHandler::AddPlayer(std::vector<GLPlayer*> players)
 {
-	for (int i = 0; i < players.size(); i++)
+	for (size_t i = 0; i < players.size(); i++)
 	{
 		this->players.push_back(players.at(i));
 	}
@@ -140,7 +140,7 @@ void GLCollisionHandler::AddPlayer(std::vector<GLPlayer*> players)
 
 void GLCollisionHandler::RemovePlayer(GLPlayer * player)
 {
-	for (int i = 0; i < this->players.size(); i++)
+	for (size_t i = 0; i < this->players.size(); i++)
 	{
 		if (this->players.at(i) == player)
 		{
@@ -157,7 +157,7 @@ void GLCollisionHandler::AddNPC(GLNPC * npc)
 
 void GLCollisionHandler::AddNPC(std::vector<GLNPC*> npcs)
 {
-	for (int i = 0; i < npcs.size(); i++)
+	for (size_t i = 0; i < npcs.size(); i++)
 	{
 		this->NPCs.push_back(npcs.at(i));
 	}
@@ -165,7 +165,7 @@ void GLCollisionHandler::AddNPC(std::vector<GLNPC*> npcs)
 
 void GLCollisionHandler::RemoveNPC(GLNPC * npc)
 {
-	for (int i = 0; i < this->NPCs.size(); i++)
+	for (size_t i = 0; i < this->NPCs.size(); i++)
 	{
 		if (this->NPCs.at(i) == npc)
 		{
@@ -181,7 +181,7 @@ void GLCollisionHandler::AddModel(GLModel * model)
 
 void GLCollisionHandler::AddModel(std::vector<GLModel*> models)
 {
-	for (int i = 0; i < models.size(); i++)
+	for (size_t i = 0; i < models.size(); i++)
 	{
 		this->models.push_back(models.at(i));
 	}
@@ -189,7 +189,7 @@ void GLCollisionHandler::AddModel(std::vector<GLModel*> models)
 
 void GLCollisionHandler::RemoveModel(GLModel * model)
 {
-	for (int i = 0; i < this->models.size(); i++)
+	for (size_t i = 0; i < this->models.size(); i++)
 	{
 		if (this->models.at(i) == model)
 		{
