+++ ProjectSDL2.1/GLNPC_GoldFish.cpp
@@ -7,7 +7,7 @@ GLNPC_GoldFish::GLNPC_GoldFish(FishBox * FSH_Loader, unsigned int modelID) : GLN
 	this->currentState = NPC_MOVE;
 	transform->SetPos(glm::vec3(RNG::range(-DEADZONEX, DEADZONEX), RNG::range(-DEADZONEY, DEADZONEY), RNG::range(-DEADZONEZ, DEADZONEZ)));
 
-	this->TimeUntilChange = RNG::range(0.2, 3.0f);
+	this->TimeUntilChange = RNG::range(0.2f, 3.0f);
 	this->scaleChange = RNG::range(-0.02f, 0.02f);
 	this->forwardSpeed = RNG::range(0.0f, 3.3f); 
 	this->rotationChange = glm::vec3(0, RNG::range(-1.0f, 1.0f), 0);
@@ -20,7 +20,7 @@ void GLNPC_GoldFish::NPCUpdate(float deltaTime)
 		TimeUntilChange -= deltaTime;
 		if (TimeUntilChange < 0)
 		{
-			this->TimeUntilChange = RNG::range(0.2, 3.0f);
+			this->TimeUntilChange = RNG::range(0.2f, 3.0f);
 			this->forwardSpeed = RNG::range(0.5f, 6.3f);
 			this->rotationChange = glm::vec3(RNG::range(-0.1f, 0.1f), RNG::range(-0.8f, 0.8f), 0);
 		}
@@ -93,7 +93,7 @@ void GLNPC_GoldFish::initiateFleeingState(glm::vec3 playerForwardVector)
 	{
 		this->currentState = NPC_FLEEING;
 		//this->transform->SetRot(playerForwardVector);
-		fleeingTimer = RNG::range(3, 10);
+		fleeingTimer = (float)RNG::range(3, 10);
 	}
 }
 
@@ -116,7 +116,7 @@ void GLNPC_GoldFish::checkboarderCollision()
 	}
 	if (pos.x < -this->DEADZONEX)
 	{
-		pos.x = -DEADZONEX;
+		pos.x = (float)-DEADZONEX;
 		if (rot.y >= 0)
 		{
 			this->rotationChange = glm::vec3(RNG::range(-0.1f, 0.1f), RNG::range(1.5f, 2.0f), 0);
@@ -128,7 +128,7 @@ void GLNPC_GoldFish::checkboarderCollision()
 	}
 	else if (pos.x >this->DEADZONEX)
 	{
-		pos.x = DEADZONEX;
+		pos.x = (float)DEADZONEX;
 		if (rot.y >= 0)
 		{
 			this->rotationChange = glm::vec3(RNG::range(-0.1f, 0.1f), RNG::range(1.5f, 2.0f), 0);
@@ -141,18 +141,18 @@ void GLNPC_GoldFish::checkboarderCollision()
 
 	if (pos.y < -this->DEADZONEY)
 	{
-		pos.y = -DEADZONEY;
+		pos.y = (float)-DEADZONEY;
 		this->rotationChange = glm::vec3(RNG::range(-0.3f, -0.2f), RNG::range(-0.8f, 0.8f), 0);
 	}
 	else if (pos.y >this->DEADZONEY)
 	{
-		pos.y = DEADZONEY;
+		pos.y = (float)DEADZONEY;
 		this->rotationChange = glm::vec3(RNG::range(0.2f, 0.3f), RNG::range(-0.8f, 0.8f), 0);
 	}
 
 	if (pos.z < -this->DEADZONEZ)
 	{
-		pos.z = -DEADZONEZ;
+		pos.z = (float)-DEADZONEZ;
 
 		if (rot.y >= 0)
 		{
@@ -165,7 +165,7 @@ void GLNPC_GoldFish::checkboarderCollision()
 	}
 	else if (pos.z >this->DEADZONEZ)
 	{
-		pos.z = DEADZONEZ;
+		pos.z = (float)DEADZONEZ;
 		
 		if (rot.y >= 0)
 		{
@@ -191,7 +191,7 @@ void GLNPC_GoldFish::ResetFish()
 	this->currentState = NPC_MOVE;
 	transform->SetPos(glm::vec3(RNG::range(-DEADZONEX, DEADZONEX), RNG::range(-DEADZONEY, DEADZONEY), RNG::range(-DEADZONEZ, DEADZONEZ)));
 
-	this->TimeUntilChange = RNG::range(0.2, 3.0f);
+	this->TimeUntilChange = RNG::range(0.2f, 3.0f);
 	this->scaleChange = RNG::range(-0.02f, 0.02f);
 	this->forwardSpeed = RNG::range(0.0f, 3.3f);
 	this->rotationChange = glm::vec3(0, RNG::range(-1.0f, 1.0f), 0);
